refactor(guard): unify custom-key check engine
- Three mirrored _check_*_custom_keys drifted-copy risks grew with each
  new write axis (update_test_runs et al. incoming).
- Add _custom_keys.check_custom_keys over axis closures; axis modules
  keep their sampling fetchers and exact error texts; drop _common.py.
